ch-setup-db_add-seeds

+++ src/database/migrations/20191027031414-create-order_products.js
@@ -31,9 +31,10 @@ module.exports = {
         allowNull: false,
         onUpdate: 'CASCADE',
         onDelete: 'NO ACTION'
+
       },
       quantity: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.INTEGER,
         allowNull: false
       }
     }
